integration: move stuff to package
