+++ integration/io_test.go
@@ -8,17 +8,19 @@ package integration
 
 import (
 	"testing"
+
+	"github.com/u-root/u-root/pkg/vmtest"
 )
 
 // TestIO tests the string "UART TEST" is written to the serial port on 0x3f8.
 func TestIO(t *testing.T) {
 	// TODO: support arm
-	if TestArch() != "amd64" {
-		t.Skipf("test not supported on %s", TestArch())
+	if vmtest.TestArch() != "amd64" {
+		t.Skipf("test not supported on %s", vmtest.TestArch())
 	}
 
 	// Create the CPIO and start QEMU.
-	q, cleanup := QEMUTest(t, &Options{
+	q, cleanup := vmtest.QEMUTest(t, &vmtest.Options{
 		Cmds: []string{
 			"github.com/u-root/u-root/integration/testcmd/io/uinit",
 			"github.com/u-root/u-root/cmds/core/init",
