+++ integration/kexec_test.go
@@ -6,17 +6,19 @@ package integration
 
 import (
 	"testing"
+
+	"github.com/u-root/u-root/pkg/vmtest"
 )
 
 // TestMountKexec runs an init which mounts a filesystem and kexecs a kernel.
 func TestMountKexec(t *testing.T) {
 	// TODO: support arm
-	if TestArch() != "amd64" {
-		t.Skipf("test not supported on %s", TestArch())
+	if vmtest.TestArch() != "amd64" {
+		t.Skipf("test not supported on %s", vmtest.TestArch())
 	}
 
 	// Create the CPIO and start QEMU.
-	q, cleanup := QEMUTest(t, &Options{
+	q, cleanup := vmtest.QEMUTest(t, &vmtest.Options{
 		Cmds: []string{
 			"github.com/u-root/u-root/integration/testcmd/kexec/uinit",
 			"github.com/u-root/u-root/cmds/core/init",
