+++ integration/multiboot_test.go
@@ -13,6 +13,7 @@ import (
 	"testing"
 
 	"github.com/u-root/u-root/pkg/multiboot"
+	"github.com/u-root/u-root/pkg/vmtest"
 )
 
 func testMultiboot(t *testing.T, kernel string) {
@@ -23,7 +24,7 @@ func testMultiboot(t *testing.T, kernel string) {
 		t.Skip("multiboot kernel is not present")
 	}
 
-	q, cleanup := QEMUTest(t, &Options{
+	q, cleanup := vmtest.QEMUTest(t, &vmtest.Options{
 		Files: []string{
 			src + ":kernel",
 		},
@@ -74,8 +75,8 @@ func testMultiboot(t *testing.T, kernel string) {
 
 func TestMultiboot(t *testing.T) {
 	// TODO: support arm
-	if TestArch() != "amd64" {
-		t.Skipf("test not supported on %s", TestArch())
+	if vmtest.TestArch() != "amd64" {
+		t.Skipf("test not supported on %s", vmtest.TestArch())
 	}
 
 	for _, kernel := range []string{"/kernel", "/kernel.gz"} {
