+++ pkg/vmtest/fields.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package integration
+package vmtest
 
 func isWhitespace(b byte) bool {
 	return b == '\t' || b == '\n' || b == '\v' ||
